+++ src/main/java/com/gamzabat/algohub/feature/group/ranking/service/RankingService.java
@@ -9,9 +9,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.gamzabat.algohub.feature.group.ranking.domain.Ranking;
 import com.gamzabat.algohub.feature.group.ranking.dto.GetRankingResponse;
-import com.gamzabat.algohub.feature.group.ranking.exception.CannotFoundRankingException;
 import com.gamzabat.algohub.feature.group.ranking.repository.RankingRepository;
-import com.gamzabat.algohub.feature.group.studygroup.domain.GroupMember;
 import com.gamzabat.algohub.feature.group.studygroup.domain.StudyGroup;
 import com.gamzabat.algohub.feature.group.studygroup.exception.CannotFoundGroupException;
 import com.gamzabat.algohub.feature.group.studygroup.exception.GroupMemberValidationException;
@@ -77,35 +75,5 @@ public class RankingService {
 				(long)r.getSolvedCount(),
 				r.getRankDiff()))
 			.toList();
-	}
-
-	public void updateRanking(GroupMember member) {
-		Ranking ranking = rankingRepository.findByMember(member)
-			.orElseThrow(() -> new CannotFoundRankingException("유저의 랭킹 정보를 조회할 수 없습니다."));
-		ranking.increaseSolvedCount();
-
-		List<Ranking> rankings = rankingRepository.findAllByStudyGroup(member.getStudyGroup());
-		rankings.sort((r1, r2) -> {
-			int compare = r2.getSolvedCount() - r1.getSolvedCount();
-			if (compare == 0)
-				return r1.getMember().getJoinDate().compareTo(r2.getMember().getJoinDate());
-			return r2.getSolvedCount() - r1.getSolvedCount();
-		});
-		for (Ranking r : rankings) {
-			int originRank = r.getCurrentRank();
-			int newRank = rankings.indexOf(r) + 1;
-			r.updateRank(newRank);
-			r.updateRankDiff(generateRankDiffString(originRank, newRank));
-		}
-	}
-
-	private String generateRankDiffString(int originRank, int newRank) {
-		int rankDiff = originRank - newRank;
-		if (rankDiff > 0)
-			return "+" + rankDiff;
-		else if (rankDiff < 0)
-			return String.valueOf(rankDiff);
-		else
-			return "-";
 	}
 }
